Separate schemas into separate typeDefs/resolvers and stitch them together
