+++ package.json
@@ -24,6 +24,9 @@
   "homepage": "https://github.com/EmilK15/node-graphql-2022#readme",
   "dependencies": {
     "@apollo/server": "^4.3.0",
+    "@graphql-tools/load-files": "^6.6.1",
+    "@graphql-tools/merge": "^8.3.14",
+    "@graphql-tools/stitch": "^8.7.33",
     "body-parser": "^1.20.1",
     "cors": "^2.8.5",
     "express": "^4.18.2",
