+++ schema/propertyOwners/index.js
@@ -1,0 +1,7 @@
+const { resolvers } = require('./resolver');
+const { typeDefs } = require('./typeDef');
+
+module.exports = {
+    resolvers,
+    typeDefs
+}
